requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pyjwt
passlib
python-dotenv
openai
python-decouple




google-generativeai
python-jose[cryptography]
passlib[bcrypt]
streamlit
requests
pandas